Remove unnecessary `toInt()` in fallback baseline calculation https://bugs.webkit.org/show_bug.cgi?id=303062 rdar://165366161
Reviewed by NOBODY (OOPS!).

RenderMathMLBlock::ascentForChild() previously used child.logicalHeight().toInt()
as a fallback when firstLineBaseline() was missing. Since LayoutUnit can be
returned directly and callers expect a LayoutUnit, the integer conversion is
unnecessary and loses precision.

This patch removes the toInt() call and returns child.logicalHeight()
directly, preserving full LayoutUnit precision and avoiding an unnecessary
fixed-point to integer conversion.

* Source/WebCore/rendering/mathml/RenderMathMLBlockInlines.h:
(WebCore::RenderMathMLBlock::ascentForChild):

* LayoutTests/imported/w3c/web-platform-tests/mathml/relations/html5-tree/dynamic-childlist-002-expected.txt: Rebaseline

+++ Source/WebCore/rendering/mathml/RenderMathMLBlockInlines.h
@@ -44,7 +44,7 @@ inline RenderMathMLTable::RenderMathMLTable(MathMLElement& element, RenderStyle&
 
 inline LayoutUnit RenderMathMLBlock::ascentForChild(const RenderBox& child)
 {
-    return child.firstLineBaseline().value_or(child.logicalHeight().toInt());
+    return child.firstLineBaseline().value_or(child.logicalHeight());
 }
 
 inline LayoutUnit RenderMathMLBlock::mirrorIfNeeded(LayoutUnit horizontalOffset, const RenderBox& child) const
